+++ packages/nutui-taro-demo/src/dentry/pages/checkbox/index.vue
@@ -227,8 +227,6 @@ export default {
   flex-wrap: wrap;
 }
 .nut-checkbox {
-  display: flex;
-  margin-right: 20px;
   .nut-checkbox__label {
     margin-left: 10px;
   }